test(local-ci): split video-setup tests under the 1k-LOC ceiling (Phase 5 code-health)
Split test_desktop_video_setup_commands_cli.py (1185 LOC) into video-doctor
cases (490) + video-setup-steps cases (831), both <1k.

Skill-Update: skip skill=ci reason="video-proof reland edits tools/local-ci and tools/cli mapped paths but does not change CI or CLI-maintenance behavior; the video-proof SKILL.md is updated"
Skill-Update: skip skill=cli-maintenance reason="video-proof reland edits tools/local-ci and tools/cli mapped paths but does not change CI or CLI-maintenance behavior; the video-proof SKILL.md is updated"
