+++ tools/local-ci/test_desktop_video_setup_steps.py
@@ -1,0 +1,831 @@
+#!/usr/bin/env python3
+"""Tests for the desktop video-doctor / video-setup commands."""
+
+from __future__ import annotations
+
+import json
+import os
+from argparse import Namespace
+from pathlib import Path
+import subprocess
+import tempfile
+import unittest
+
+from module_test_utils import load_local_ci_module
+
+
+def load_module():
+    return load_local_ci_module("desktop_video_setup_commands_cli.py", add_module_dir=True)
+
+
+class DesktopVideoSetupStepsTests(unittest.TestCase):
+    def setUp(self):
+        self.mod = load_module()
+        self.printed: list[str] = []
+        self.writes: list[tuple[Path, str]] = []
+        self.saved_configs: list[dict] = []
+        self.targets = {
+            "mac": {
+                "adapter": "macos-local",
+                "bootstrap": "launchagent",
+                "target_type": "local",
+                "capability_tier": "full",
+            },
+            "ubuntu": {
+                "adapter": "linux-xvfb",
+                "bootstrap": "xvfb-run",
+                "target_type": "ssh",
+                "capability_tier": "full",
+                "host": "ubuntu",
+            },
+            "windows": {
+                "adapter": "windows-session-agent",
+                "bootstrap": "scheduled-task",
+                "target_type": "ssh",
+                "capability_tier": "full",
+                "host": "win",
+                "repo_path": r"C:\Old\Pulp",
+            },
+        }
+
+    def print_line(self, line: str):
+        self.printed.append(line)
+
+    def config(self):
+        return {
+            "defaults": {},
+            "desktop_automation": {
+                "artifact_root": "/tmp/desktop-artifacts",
+                "targets": self.targets,
+            },
+        }
+
+    def setup_ok_checks(self):
+        return [
+            {"name": "setup.pulp", "ok": True, "detail": "/usr/local/bin/pulp", "required": True, "title": "Pulp CLI"},
+            {"name": "setup.npm", "ok": True, "detail": "/usr/local/bin/npm", "required": True, "title": "npm"},
+            {"name": "setup.node", "ok": True, "detail": "/usr/local/bin/node", "required": True, "title": "Node.js"},
+            {"name": "setup.cmake", "ok": True, "detail": "/usr/local/bin/cmake", "required": True, "title": "CMake"},
+        ]
+
+    def deps(self):
+        def resolve(_config, name):
+            return self.targets[name]
+
+        def update_repo_path(config, name, repo_path):
+            config["desktop_automation"]["targets"][name]["repo_path"] = repo_path
+            self.targets[name]["repo_path"] = repo_path
+
+        return {
+            "load_config_fn": self.config,
+            "resolve_desktop_target_fn": resolve,
+            "check_writable_dir_fn": lambda _path: (True, ""),
+            "desktop_target_contract_fn": lambda name, target: {
+                "target": name,
+                "task_name": f"PulpDesktopAutomationAgent-{name}",
+                "remote_root": r"C:\Users\daniel\AppData\Local\Pulp\desktop-automation-agent"
+                if target["adapter"] == "windows-session-agent"
+                else None,
+            },
+            "ensure_host_reachable_fn": lambda _name, target, _defaults: target.get("host"),
+            "bootstrap_windows_session_agent_fn": lambda _host, _contract: {
+                "remote_root": r"C:\RemoteRoot",
+                "script_path": r"C:\RemoteRoot\agent.ps1",
+            },
+            "probe_windows_session_agent_fn": lambda _host, _contract: {
+                "task_present": True,
+                "agent_root_exists": True,
+                "jobs_dir_exists": True,
+                "results_dir_exists": True,
+                "script_exists": True,
+            },
+            "subprocess_run_fn": lambda *_args, **_kwargs: subprocess.CompletedProcess(["git"], 0, "a" * 40 + "\n", ""),
+            "root_path": Path("/repo"),
+            "new_install_job_id_fn": lambda: "install123",
+            "sync_job_bundle_to_ssh_host_fn": lambda _host, job: (f"{job['id']}.git", "refs/pulp-ci-bundles/install"),
+            "ensure_windows_remote_tooling_fn": lambda _host: {
+                "probe": {
+                    "git_found": True,
+                    "git_path": r"C:\Program Files\Git\cmd\git.exe",
+                    "git_version": "git version 2.49.0.windows.1",
+                },
+                "installed": ["git"],
+            },
+            "windows_remote_tooling_ready_fn": lambda probe: bool(probe.get("git_found")),
+            "ensure_windows_remote_repo_checkout_fn": lambda *_args, **_kwargs: {
+                "repo_path": r"C:\Users\daniel\pulp-validate",
+                "repo_exists": True,
+            },
+            "git_origin_clone_url_fn": lambda _root: "https://github.com/danielraffel/pulp",
+            "windows_repo_checkout_ready_fn": lambda probe: bool(probe.get("repo_exists")),
+            "update_target_repo_path_fn": update_repo_path,
+            "save_config_fn": self.saved_configs.append,
+            "now_iso_fn": lambda: "2026-06-11T00:00:00Z",
+            "desktop_target_receipt_path_fn": lambda name: Path(f"/receipts/{name}.json"),
+            "atomic_write_text_fn": lambda path, text: self.writes.append((path, text)),
+            "windows_tooling_detail_fn": lambda probe, tool_name, **_kwargs: f"{probe.get(tool_name + '_version')} ({probe.get(tool_name + '_path')})",
+            "print_fn": self.print_line,
+        }
+
+    def written_receipt(self):
+        self.assertEqual(len(self.writes), 1)
+        return json.loads(self.writes[0][1])
+
+    def test_video_setup_prints_portable_first_run_steps(self):
+        deps = {
+            "load_config_fn": self.config,
+            "resolve_desktop_target_fn": lambda _config, name: self.targets[name],
+            "desktop_doctor_checks_fn": lambda *_args: self.fail("doctor should not run without --check"),
+            "normalize_desktop_optional_config_fn": lambda optional: {"video_capture": bool((optional or {}).get("video_capture"))},
+            "video_proof_smoke_fn": lambda: self.fail("smoke should not run without --check"),
+            "probe_macos_avfoundation_audio_fn": lambda _device: self.fail("audio probe should not run without --check"),
+            "setup_prerequisite_checks_fn": self.setup_ok_checks,
+            "print_fn": self.print_line,
+        }
+
+        result = self.mod.cmd_desktop_video_setup(
+            Namespace(target="mac", machine="blackbook", check=False, skip_remotion_smoke=False, video_audio="none", video_audio_device=None, json=False),
+            **deps,
+        )
+
+        self.assertEqual(result, 0)
+        self.assertIn("Desktop video setup for `mac`", self.printed)
+        self.assertIn("  machine: blackbook", self.printed)
+        self.assertTrue(any("cp tools/local-ci/config.example.json tools/local-ci/config.json" in line for line in self.printed))
+        self.assertTrue(any("npm --prefix tools/local-ci install" in line for line in self.printed))
+        self.assertTrue(any("future: pulp tool install video-proof" in line for line in self.printed))
+        self.assertTrue(any("pulp tool info video-proof --json" in line for line in self.printed))
+        self.assertTrue(any("pulp tool doctor video-proof --run" in line for line in self.printed))
+        self.assertTrue(any("target.mac.video_capture true" in line for line in self.printed))
+        self.assertTrue(any("--video-audio system" in line for line in self.printed))
+        self.assertTrue(any("--label blackbook-video-setup-smoke" in line for line in self.printed))
+
+    def test_video_setup_prints_first_run_steps_without_config(self):
+        deps = {
+            "load_config_fn": lambda: (_ for _ in ()).throw(FileNotFoundError("Local CI config not found at /repo/tools/local-ci/config.json")),
+            "resolve_desktop_target_fn": lambda *_args: self.fail("target should not be resolved without config"),
+            "desktop_doctor_checks_fn": lambda *_args: self.fail("doctor should not run without --check"),
+            "normalize_desktop_optional_config_fn": lambda optional: {"video_capture": bool((optional or {}).get("video_capture"))},
+            "video_proof_smoke_fn": lambda: self.fail("smoke should not run without --check"),
+            "probe_macos_avfoundation_audio_fn": lambda _device: self.fail("audio probe should not run without --check"),
+            "setup_prerequisite_checks_fn": self.setup_ok_checks,
+            "print_fn": self.print_line,
+        }
+
+        result = self.mod.cmd_desktop_video_setup(
+            Namespace(target="mac", machine="blackbook", check=False, skip_remotion_smoke=False, video_audio="none", video_audio_device=None, json=False),
+            **deps,
+        )
+
+        self.assertEqual(result, 0)
+        self.assertIn("  adapter: macos-local", self.printed)
+        self.assertTrue(any("cp tools/local-ci/config.example.json tools/local-ci/config.json" in line for line in self.printed))
+
+    def test_video_setup_json_reports_missing_config_during_check(self):
+        deps = {
+            "load_config_fn": lambda: (_ for _ in ()).throw(FileNotFoundError("Local CI config not found at /repo/tools/local-ci/config.json")),
+            "resolve_desktop_target_fn": lambda *_args: self.fail("target should not be resolved without config"),
+            "desktop_doctor_checks_fn": lambda *_args: self.fail("doctor should not run without config"),
+            "normalize_desktop_optional_config_fn": lambda optional: {"video_capture": bool((optional or {}).get("video_capture"))},
+            "video_proof_smoke_fn": lambda: self.fail("smoke should not run without config"),
+            "probe_macos_avfoundation_audio_fn": lambda _device: self.fail("audio probe should not run without config"),
+            "setup_prerequisite_checks_fn": self.setup_ok_checks,
+            "print_fn": self.print_line,
+        }
+
+        result = self.mod.cmd_desktop_video_setup(
+            Namespace(target="mac", machine="blackbook", check=True, skip_remotion_smoke=False, video_audio="none", video_audio_device=None, json=True),
+            **deps,
+        )
+
+        self.assertEqual(result, 1)
+        payload = json.loads(self.printed[0])
+        self.assertFalse(payload["check"]["ok"])
+        self.assertEqual(payload["install_model"]["package_format"], "not_pulp_add")
+        self.assertEqual(payload["install_model"]["pack_manifest_schema"], "pulp.video-proof-tool-package.v1")
+        self.assertTrue(payload["setup_prerequisites"]["ok"])
+        prereq_checks_by_name = {check["name"]: check for check in payload["setup_prerequisites"]["checks"]}
+        self.assertEqual(prereq_checks_by_name["setup.pulp"]["detail"], "/usr/local/bin/pulp")
+        self.assertIsNone(payload["tool_addon"])
+        self.assertEqual(payload["check"]["checks"][0]["name"], "config")
+        self.assertEqual(payload["check"]["remediations"][0]["command"], "cp tools/local-ci/config.example.json tools/local-ci/config.json")
+        self.assertEqual(payload["steps"][0]["name"], "create_config")
+
+    def test_video_setup_json_can_check_tool_addon_without_config(self):
+        tool_addon_calls = []
+
+        def fake_tool_addon_checks(**kwargs):
+            tool_addon_calls.append(kwargs.get("pulp_command"))
+            return [
+                {"name": "tool_addon.info", "ok": True, "detail": "video-proof scope=machine lane=tool_addon format=not_pulp_add", "required": True},
+                {"name": "tool_addon.doctor", "ok": True, "detail": "smoke ok", "required": True},
+            ]
+
+        deps = {
+            "load_config_fn": lambda: (_ for _ in ()).throw(FileNotFoundError("Local CI config not found at /repo/tools/local-ci/config.json")),
+            "resolve_desktop_target_fn": lambda *_args: self.fail("target should not be resolved without config"),
+            "desktop_doctor_checks_fn": lambda *_args: self.fail("doctor should not run without config"),
+            "normalize_desktop_optional_config_fn": lambda optional: {"video_capture": bool((optional or {}).get("video_capture"))},
+            "video_proof_smoke_fn": lambda: self.fail("smoke should not run without config"),
+            "probe_macos_avfoundation_audio_fn": lambda _device: self.fail("audio probe should not run without config"),
+            "setup_prerequisite_checks_fn": self.setup_ok_checks,
+            "tool_addon_checks_fn": fake_tool_addon_checks,
+            "print_fn": self.print_line,
+        }
+
+        result = self.mod.cmd_desktop_video_setup(
+            Namespace(
+                target="mac",
+                machine="blackbook",
+                check=True,
+                check_tool_addon=True,
+                pulp_command="./build-video-proof-cli/tools/cli/pulp-cpp",
+                skip_remotion_smoke=False,
+                video_audio="none",
+                video_audio_device=None,
+                json=True,
+            ),
+            **deps,
+        )
+
+        self.assertEqual(result, 1)
+        self.assertEqual(tool_addon_calls, ["./build-video-proof-cli/tools/cli/pulp-cpp"])
+        payload = json.loads(self.printed[0])
+        self.assertFalse(payload["check"]["ok"])
+        self.assertTrue(payload["setup_prerequisites"]["ok"])
+        self.assertTrue(payload["tool_addon"]["ok"])
+        self.assertEqual(
+            payload["install_model"]["tool_info_command"],
+            "./build-video-proof-cli/tools/cli/pulp-cpp tool info video-proof --json",
+        )
+        self.assertEqual(
+            payload["install_model"]["future_command"],
+            "./build-video-proof-cli/tools/cli/pulp-cpp tool install video-proof",
+        )
+        self.assertEqual(
+            payload["install_model"]["future_check_command"],
+            "./build-video-proof-cli/tools/cli/pulp-cpp tool doctor video-proof --run",
+        )
+        self.assertEqual(
+            payload["steps"][2]["command"],
+            "./build-video-proof-cli/tools/cli/pulp-cpp tool info video-proof --json",
+        )
+        self.assertEqual(
+            payload["steps"][3]["command"],
+            "./build-video-proof-cli/tools/cli/pulp-cpp tool doctor video-proof --run",
+        )
+        self.assertEqual(payload["tool_addon"]["remediations"], [])
+        self.assertEqual(payload["check"]["checks"][0]["name"], "config")
+
+    def test_video_setup_init_config_helper_creates_without_overwriting(self):
+        with tempfile.TemporaryDirectory() as tmp:
+            root = Path(tmp)
+            example = root / "config.example.json"
+            destination = root / "config.json"
+            example.write_text('{"desktop_automation": {"targets": {}}}\n')
+
+            created = self.mod.desktop_video_init_config(config_path=destination, example_path=example)
+            self.assertTrue(created["ok"])
+            self.assertTrue(created["created"])
+            self.assertEqual(destination.read_text(), example.read_text())
+
+            destination.write_text('{"custom": true}\n')
+            existing = self.mod.desktop_video_init_config(config_path=destination, example_path=example)
+            self.assertTrue(existing["ok"])
+            self.assertFalse(existing["created"])
+            self.assertEqual(destination.read_text(), '{"custom": true}\n')
+
+    def test_video_setup_init_config_helper_honors_env_config_path(self):
+        with tempfile.TemporaryDirectory() as tmp:
+            root = Path(tmp)
+            example = root / "config.example.json"
+            destination = root / "nested" / "config.json"
+            example.write_text('{"desktop_automation": {"targets": {}}}\n')
+            previous = os.environ.get("PULP_LOCAL_CI_CONFIG")
+            os.environ["PULP_LOCAL_CI_CONFIG"] = str(destination)
+            try:
+                created = self.mod.desktop_video_init_config(example_path=example)
+            finally:
+                if previous is None:
+                    os.environ.pop("PULP_LOCAL_CI_CONFIG", None)
+                else:
+                    os.environ["PULP_LOCAL_CI_CONFIG"] = previous
+
+            self.assertTrue(created["ok"])
+            self.assertTrue(created["created"])
+            self.assertEqual(created["path"], str(destination))
+            self.assertEqual(destination.read_text(), example.read_text())
+
+    def test_video_setup_enable_target_capture_helper(self):
+        config = {"desktop_automation": {"targets": {"mac": {"optional": {"video_capture": False}}}}}
+
+        payload = self.mod.desktop_video_enable_target_capture(config, "mac")
+
+        self.assertTrue(payload["ok"])
+        self.assertTrue(payload["changed"])
+        self.assertEqual(payload["field"], "target.mac.video_capture")
+        self.assertTrue(config["desktop_automation"]["targets"]["mac"]["optional"]["video_capture"])
+
+    def test_video_setup_init_config_then_runs_check(self):
+        self.targets["mac"]["optional"] = {"video_capture": True}
+        calls = {"load": 0, "init": 0}
+        checks = [
+            {"name": "receipt", "ok": True, "detail": "installed"},
+            {"name": "screencapture", "ok": True, "detail": "/usr/sbin/screencapture"},
+            {"name": "video_capture", "ok": True, "detail": "/repo/node_modules/ffmpeg-static/ffmpeg", "required": False},
+            {"name": "avfoundation_screen", "ok": True, "detail": "Capture screen 0 (3:)", "required": False},
+        ]
+
+        def load_config_once_missing():
+            calls["load"] += 1
+            if calls["init"] == 0:
+                raise FileNotFoundError("Local CI config not found")
+            return self.config()
+
+        def init_config():
+            calls["init"] += 1
+            return {"ok": True, "created": True, "path": "/repo/tools/local-ci/config.json", "source": "/repo/tools/local-ci/config.example.json", "detail": "created config"}
+
+        deps = {
+            "load_config_fn": load_config_once_missing,
+            "resolve_desktop_target_fn": lambda _config, name: self.targets[name],
+            "desktop_doctor_checks_fn": lambda _config, _name: [dict(check) for check in checks],
+            "normalize_desktop_optional_config_fn": lambda optional: {"video_capture": bool((optional or {}).get("video_capture"))},
+            "video_proof_smoke_fn": lambda: {"ok": True, "detail": "smoke ok"},
+            "probe_macos_avfoundation_audio_fn": lambda _device: self.fail("audio probe should not run"),
+            "setup_prerequisite_checks_fn": self.setup_ok_checks,
+            "init_config_fn": init_config,
+            "print_fn": self.print_line,
+        }
+
+        result = self.mod.cmd_desktop_video_setup(
+            Namespace(
+                target="mac",
+                machine="blackbook",
+                init_config=True,
+                enable_video_capture=False,
+                check=True,
+                check_tool_addon=False,
+                skip_remotion_smoke=False,
+                video_audio="none",
+                video_audio_device=None,
+                json=True,
+            ),
+            **deps,
+        )
+
+        self.assertEqual(result, 0)
+        self.assertEqual(calls, {"load": 1, "init": 1})
+        payload = json.loads(self.printed[0])
+        self.assertTrue(payload["init_config"]["created"])
+        self.assertTrue(payload["setup_prerequisites"]["ok"])
+        self.assertTrue(payload["check"]["ok"])
+
+    def test_video_setup_can_enable_video_capture_before_check(self):
+        self.targets["mac"]["optional"] = {"video_capture": False}
+        saved_configs = []
+        checks = [
+            {"name": "receipt", "ok": True, "detail": "installed"},
+            {"name": "screencapture", "ok": True, "detail": "/usr/sbin/screencapture"},
+            {"name": "video_capture", "ok": True, "detail": "/repo/node_modules/ffmpeg-static/ffmpeg", "required": False},
+            {"name": "avfoundation_screen", "ok": True, "detail": "Capture screen 0 (3:)", "required": False},
+        ]
+        config = self.config()
+        config["desktop_automation"]["targets"]["mac"]["optional"]["video_capture"] = False
+
+        def resolve(_config, name):
+            self.targets[name] = _config["desktop_automation"]["targets"][name]
+            return self.targets[name]
+
+        deps = {
+            "load_config_fn": lambda: config,
+            "save_config_fn": saved_configs.append,
+            "resolve_desktop_target_fn": resolve,
+            "desktop_doctor_checks_fn": lambda _config, _name: [dict(check) for check in checks],
+            "normalize_desktop_optional_config_fn": lambda optional: {"video_capture": bool((optional or {}).get("video_capture"))},
+            "video_proof_smoke_fn": lambda: {"ok": True, "detail": "smoke ok"},
+            "probe_macos_avfoundation_audio_fn": lambda _device: self.fail("audio probe should not run"),
+            "setup_prerequisite_checks_fn": self.setup_ok_checks,
+            "print_fn": self.print_line,
+        }
+
+        result = self.mod.cmd_desktop_video_setup(
+            Namespace(
+                target="mac",
+                machine="blackbook",
+                init_config=False,
+                enable_video_capture=True,
+                check=True,
+                check_tool_addon=False,
+                skip_remotion_smoke=False,
+                video_audio="none",
+                video_audio_device=None,
+                json=True,
+            ),
+            **deps,
+        )
+
+        self.assertEqual(result, 0)
+        self.assertEqual(len(saved_configs), 1)
+        self.assertTrue(saved_configs[0]["desktop_automation"]["targets"]["mac"]["optional"]["video_capture"])
+        payload = json.loads(self.printed[0])
+        self.assertTrue(payload["target_config"]["changed"])
+        self.assertTrue(payload["check"]["ok"])
+        checks_by_name = {check["name"]: check for check in payload["check"]["checks"]}
+        self.assertTrue(checks_by_name["target.video_capture"]["ok"])
+
+    def test_video_setup_text_reports_setup_and_tool_addon_without_config(self):
+        deps = {
+            "load_config_fn": lambda: (_ for _ in ()).throw(FileNotFoundError("Local CI config not found at /repo/tools/local-ci/config.json")),
+            "resolve_desktop_target_fn": lambda *_args: self.fail("target should not be resolved without config"),
+            "desktop_doctor_checks_fn": lambda *_args: self.fail("doctor should not run without config"),
+            "normalize_desktop_optional_config_fn": lambda optional: {"video_capture": bool((optional or {}).get("video_capture"))},
+            "video_proof_smoke_fn": lambda: self.fail("smoke should not run without config"),
+            "probe_macos_avfoundation_audio_fn": lambda _device: self.fail("audio probe should not run without config"),
+            "setup_prerequisite_checks_fn": self.setup_ok_checks,
+            "tool_addon_checks_fn": lambda: [
+                {"name": "tool_addon.info", "ok": False, "detail": "pulp does not support tool", "required": True},
+                {"name": "tool_addon.doctor", "ok": False, "detail": "skipped because tool info failed", "required": True},
+            ],
+            "print_fn": self.print_line,
+        }
+
+        result = self.mod.cmd_desktop_video_setup(
+            Namespace(
+                target="mac",
+                machine="blackbook",
+                init_config=False,
+                check=True,
+                check_tool_addon=True,
+                skip_remotion_smoke=False,
+                video_audio="none",
+                video_audio_device=None,
+                json=False,
+            ),
+            **deps,
+        )
+
+        self.assertEqual(result, 1)
+        self.assertIn("Setup prerequisites: PASS", self.printed)
+        self.assertIn("Tool add-on check: FAIL", self.printed)
+        self.assertTrue(any("pulp does not support tool" in line for line in self.printed))
+        self.assertIn("Current check: FAIL", self.printed)
+
+    def test_video_setup_json_can_include_current_doctor_check(self):
+        self.targets["mac"]["optional"] = {"video_capture": True}
+        matrix_calls = []
+        checks = [
+            {"name": "receipt", "ok": True, "detail": "installed"},
+            {"name": "screencapture", "ok": True, "detail": "/usr/sbin/screencapture"},
+            {"name": "video_capture", "ok": True, "detail": "/repo/node_modules/ffmpeg-static/ffmpeg", "required": False},
+            {"name": "avfoundation_screen", "ok": True, "detail": "Capture screen 0 (3:)", "required": False},
+        ]
+        deps = {
+            "load_config_fn": self.config,
+            "resolve_desktop_target_fn": lambda _config, name: self.targets[name],
+            "desktop_doctor_checks_fn": lambda _config, _name: [dict(check) for check in checks],
+            "normalize_desktop_optional_config_fn": lambda optional: {"video_capture": bool((optional or {}).get("video_capture"))},
+            "video_proof_smoke_fn": lambda: {"ok": True, "detail": "smoke ok"},
+            "probe_macos_avfoundation_audio_fn": lambda device: (device == "2", "BlackHole 2ch (2)" if device == "2" else "missing"),
+            "setup_prerequisite_checks_fn": self.setup_ok_checks,
+            "desktop_video_matrix_payload_fn": lambda **kwargs: matrix_calls.append(kwargs) or {
+                "kind": "desktop-video-proof-demo-matrix",
+                "checked": True,
+                "scenarios": [
+                    {
+                        "id": "audio-inspector-demo",
+                        "status": "ready",
+                        "declared_status": "ready",
+                    },
+                    {
+                        "id": "design-parity",
+                        "status": "blocked",
+                        "declared_status": "ready",
+                        "local_readiness": {
+                            "checks": [
+                                {
+                                    "name": "source_image",
+                                    "ok": False,
+                                    "required": True,
+                                    "detail": "missing planning/screenshots/reference.png",
+                                }
+                            ],
+                        },
+                    },
+                ],
+            },
+            "print_fn": self.print_line,
+        }
+
+        result = self.mod.cmd_desktop_video_setup(
+            Namespace(
+                target="mac",
+                machine="blackbook",
+                check=True,
+                skip_remotion_smoke=False,
+                video_audio="system",
+                video_audio_device="2",
+                design_parity_manifest="/tmp/run/manifest.json",
+                design_parity_source_image="/tmp/source.png",
+                design_parity_native_image="/tmp/native.png",
+                json=True,
+            ),
+            **deps,
+        )
+
+        self.assertEqual(result, 0)
+        payload = json.loads(self.printed[0])
+        self.assertEqual(payload["machine"], "blackbook")
+        self.assertEqual(payload["install_model"]["future_command"], "pulp tool install video-proof")
+        self.assertEqual(payload["install_model"]["future_check_command"], "pulp tool doctor video-proof --run")
+        self.assertEqual(payload["install_model"]["tool_info_command"], "pulp tool info video-proof --json")
+        self.assertEqual(payload["install_model"]["pack_command"], "python3 tools/local-ci/pack_video_proof_tool.py --json")
+        self.assertEqual(payload["install_model"]["pack_npm_script"], "npm --prefix tools/local-ci run pack-video-proof-tool -- --json")
+        self.assertEqual(payload["install_model"]["verify_command"], "python3 tools/local-ci/pack_video_proof_tool.py --verify <manifest> --json")
+        self.assertEqual(payload["install_model"]["artifact_install_command"], "pulp tool install video-proof --artifact-manifest <manifest> --force")
+        self.assertEqual(payload["install_model"]["pack_manifest_schema"], "pulp.video-proof-tool-package.v1")
+        self.assertEqual(payload["install_model"]["install_scope"], "machine")
+        self.assertEqual(payload["install_model"]["distribution_lane"], "tool_addon")
+        self.assertEqual(payload["install_model"]["package_format"], "not_pulp_add")
+        self.assertEqual(payload["install_model"]["artifact_status"], "source_tree_iteration")
+        self.assertTrue(payload["setup_prerequisites"]["ok"])
+        prereq_checks_by_name = {check["name"]: check for check in payload["setup_prerequisites"]["checks"]}
+        self.assertEqual(prereq_checks_by_name["setup.pulp"]["detail"], "/usr/local/bin/pulp")
+        self.assertTrue(payload["check"]["ok"])
+        self.assertEqual(payload["check"]["target"], "mac")
+        checks_by_name = {check["name"]: check for check in payload["check"]["checks"]}
+        self.assertTrue(checks_by_name["avfoundation_audio"]["ok"])
+        self.assertEqual(
+            matrix_calls,
+            [
+                {
+                    "target": "mac",
+                    "check": True,
+                    "design_parity_manifest": "/tmp/run/manifest.json",
+                    "design_parity_source_image": "/tmp/source.png",
+                    "design_parity_native_image": "/tmp/native.png",
+                }
+            ],
+        )
+        self.assertEqual(payload["demo_matrix"]["checked"], True)
+        self.assertEqual(payload["demo_matrix"]["scenarios"][1]["id"], "design-parity")
+        self.assertEqual(payload["demo_matrix"]["scenarios"][1]["status"], "blocked")
+        self.assertEqual(payload["demo_matrix"]["scenarios"][1]["declared_status"], "ready")
+        self.assertEqual(payload["steps"][2]["name"], "inspect_tool_addon")
+        self.assertEqual(payload["steps"][2]["command"], "pulp tool info video-proof --json")
+        self.assertEqual(payload["steps"][3]["name"], "check_tool_addon")
+        self.assertEqual(payload["steps"][3]["command"], "pulp tool doctor video-proof --run")
+        self.assertEqual(payload["steps"][8]["name"], "audio_doctor")
+        self.assertEqual(payload["steps"][9]["name"], "smoke_proof")
+        self.assertIn("--run-in-terminal", payload["steps"][7]["command"])
+
+    def test_video_setup_check_fails_when_setup_prerequisites_are_missing(self):
+        self.targets["mac"]["optional"] = {"video_capture": True}
+        checks = [
+            {"name": "receipt", "ok": True, "detail": "installed"},
+            {"name": "screencapture", "ok": True, "detail": "/usr/sbin/screencapture"},
+            {"name": "video_capture", "ok": True, "detail": "/repo/node_modules/ffmpeg-static/ffmpeg", "required": False},
+            {"name": "avfoundation_screen", "ok": True, "detail": "Capture screen 0 (3:)", "required": False},
+        ]
+        deps = {
+            "load_config_fn": self.config,
+            "resolve_desktop_target_fn": lambda _config, name: self.targets[name],
+            "desktop_doctor_checks_fn": lambda _config, _name: [dict(check) for check in checks],
+            "normalize_desktop_optional_config_fn": lambda optional: {"video_capture": bool((optional or {}).get("video_capture"))},
+            "video_proof_smoke_fn": lambda: {"ok": True, "detail": "smoke ok"},
+            "probe_macos_avfoundation_audio_fn": lambda _device: self.fail("audio probe should not run"),
+            "setup_prerequisite_checks_fn": lambda: [
+                {"name": "setup.pulp", "ok": False, "detail": "required for tool install", "required": True, "title": "Pulp CLI", "remediation": "Install Pulp."},
+                {"name": "setup.npm", "ok": False, "detail": "required for npm install", "required": True, "title": "npm", "remediation": "Install npm."},
+                {"name": "setup.node", "ok": True, "detail": "/opt/homebrew/bin/node", "required": True, "title": "Node.js"},
+                {"name": "setup.cmake", "ok": False, "detail": "required for CLI build", "required": True, "title": "CMake", "remediation": "Install CMake."},
+            ],
+            "print_fn": self.print_line,
+        }
+
+        result = self.mod.cmd_desktop_video_setup(
+            Namespace(target="mac", machine="blackbook", check=True, skip_remotion_smoke=False, video_audio="none", video_audio_device=None, json=True),
+            **deps,
+        )
+
+        self.assertEqual(result, 1)
+        payload = json.loads(self.printed[0])
+        self.assertFalse(payload["setup_prerequisites"]["ok"])
+        self.assertTrue(payload["check"]["ok"])
+        missing = [item["check"] for item in payload["setup_prerequisites"]["remediations"]]
+        self.assertEqual(missing, ["setup.pulp", "setup.npm", "setup.cmake"])
+
+    def test_video_setup_check_can_validate_tool_addon(self):
+        self.targets["mac"]["optional"] = {"video_capture": True}
+        checks = [
+            {"name": "receipt", "ok": True, "detail": "installed"},
+            {"name": "screencapture", "ok": True, "detail": "/usr/sbin/screencapture"},
+            {"name": "video_capture", "ok": True, "detail": "/repo/node_modules/ffmpeg-static/ffmpeg", "required": False},
+            {"name": "avfoundation_screen", "ok": True, "detail": "Capture screen 0 (3:)", "required": False},
+        ]
+        tool_addon_calls = []
+
+        def fake_tool_addon_checks(**kwargs):
+            tool_addon_calls.append(kwargs.get("pulp_command"))
+            return [
+                {"name": "tool_addon.info", "ok": True, "detail": "video-proof scope=machine lane=tool_addon format=not_pulp_add", "required": True},
+                {"name": "tool_addon.doctor", "ok": True, "detail": "smoke ok", "required": True},
+            ]
+
+        deps = {
+            "load_config_fn": self.config,
+            "resolve_desktop_target_fn": lambda _config, name: self.targets[name],
+            "desktop_doctor_checks_fn": lambda _config, _name: [dict(check) for check in checks],
+            "normalize_desktop_optional_config_fn": lambda optional: {"video_capture": bool((optional or {}).get("video_capture"))},
+            "video_proof_smoke_fn": lambda: {"ok": True, "detail": "smoke ok"},
+            "probe_macos_avfoundation_audio_fn": lambda _device: self.fail("audio probe should not run"),
+            "setup_prerequisite_checks_fn": self.setup_ok_checks,
+            "tool_addon_checks_fn": fake_tool_addon_checks,
+            "print_fn": self.print_line,
+        }
+
+        result = self.mod.cmd_desktop_video_setup(
+            Namespace(
+                target="mac",
+                machine="blackbook",
+                check=True,
+                check_tool_addon=True,
+                pulp_command="./build-video-proof-cli/tools/cli/pulp-cpp",
+                skip_remotion_smoke=False,
+                video_audio="none",
+                video_audio_device=None,
+                json=True,
+            ),
+            **deps,
+        )
+
+        self.assertEqual(result, 0)
+        self.assertEqual(tool_addon_calls, ["./build-video-proof-cli/tools/cli/pulp-cpp"])
+        payload = json.loads(self.printed[0])
+        self.assertTrue(payload["check"]["ok"])
+        self.assertTrue(payload["tool_addon"]["ok"])
+        tool_checks_by_name = {check["name"]: check for check in payload["tool_addon"]["checks"]}
+        self.assertEqual(tool_checks_by_name["tool_addon.info"]["detail"], "video-proof scope=machine lane=tool_addon format=not_pulp_add")
+        self.assertEqual(
+            payload["steps"][2]["command"],
+            "./build-video-proof-cli/tools/cli/pulp-cpp tool info video-proof --json",
+        )
+        self.assertEqual(
+            payload["steps"][3]["command"],
+            "./build-video-proof-cli/tools/cli/pulp-cpp tool doctor video-proof --run",
+        )
+        self.assertEqual(
+            payload["install_model"]["artifact_install_command"],
+            "./build-video-proof-cli/tools/cli/pulp-cpp tool install video-proof --artifact-manifest <manifest> --force",
+        )
+        self.assertEqual(payload["tool_addon"]["remediations"], [])
+
+    def test_video_setup_check_reports_tool_addon_remediation(self):
+        self.targets["mac"]["optional"] = {"video_capture": True}
+        checks = [
+            {"name": "receipt", "ok": True, "detail": "installed"},
+            {"name": "screencapture", "ok": True, "detail": "/usr/sbin/screencapture"},
+            {"name": "video_capture", "ok": True, "detail": "/repo/node_modules/ffmpeg-static/ffmpeg", "required": False},
+            {"name": "avfoundation_screen", "ok": True, "detail": "Capture screen 0 (3:)", "required": False},
+        ]
+        deps = {
+            "load_config_fn": self.config,
+            "resolve_desktop_target_fn": lambda _config, name: self.targets[name],
+            "desktop_doctor_checks_fn": lambda _config, _name: [dict(check) for check in checks],
+            "normalize_desktop_optional_config_fn": lambda optional: {"video_capture": bool((optional or {}).get("video_capture"))},
+            "video_proof_smoke_fn": lambda: {"ok": True, "detail": "smoke ok"},
+            "probe_macos_avfoundation_audio_fn": lambda _device: self.fail("audio probe should not run"),
+            "setup_prerequisite_checks_fn": self.setup_ok_checks,
+            "tool_addon_checks_fn": lambda: [
+                {"name": "tool_addon.info", "ok": False, "detail": "tool not installed", "required": True},
+                {"name": "tool_addon.doctor", "ok": False, "detail": "skipped because tool info failed", "required": True},
+            ],
+            "print_fn": self.print_line,
+        }
+
+        result = self.mod.cmd_desktop_video_setup(
+            Namespace(target="mac", machine="blackbook", check=True, check_tool_addon=True, skip_remotion_smoke=False, video_audio="none", video_audio_device=None, json=True),
+            **deps,
+        )
+
+        self.assertEqual(result, 1)
+        payload = json.loads(self.printed[0])
+        self.assertTrue(payload["check"]["ok"])
+        self.assertFalse(payload["tool_addon"]["ok"])
+        self.assertEqual(payload["tool_addon"]["remediations"][0]["command"], "pulp tool install video-proof")
+        self.assertEqual(payload["tool_addon"]["remediations"][0]["check_command"], "pulp tool doctor video-proof --run")
+
+    def test_video_setup_check_can_probe_remote_host_prerequisites(self):
+        self.targets["mac"]["optional"] = {"video_capture": True}
+        checks = [
+            {"name": "receipt", "ok": True, "detail": "installed"},
+            {"name": "screencapture", "ok": True, "detail": "/usr/sbin/screencapture"},
+            {"name": "video_capture", "ok": True, "detail": "/repo/node_modules/ffmpeg-static/ffmpeg", "required": False},
+            {"name": "avfoundation_screen", "ok": True, "detail": "Capture screen 0 (3:)", "required": False},
+        ]
+        probed_hosts = []
+        deps = {
+            "load_config_fn": self.config,
+            "resolve_desktop_target_fn": lambda _config, name: self.targets[name],
+            "desktop_doctor_checks_fn": lambda _config, _name: [dict(check) for check in checks],
+            "normalize_desktop_optional_config_fn": lambda optional: {"video_capture": bool((optional or {}).get("video_capture"))},
+            "video_proof_smoke_fn": lambda: {"ok": True, "detail": "smoke ok"},
+            "probe_macos_avfoundation_audio_fn": lambda _device: self.fail("audio probe should not run"),
+            "setup_prerequisite_checks_fn": self.setup_ok_checks,
+            "remote_setup_prerequisite_checks_fn": lambda host: probed_hosts.append(host) or [
+                {"name": "remote_setup.pulp", "ok": False, "detail": "missing", "required": True, "title": "Pulp CLI", "remediation": "Install Pulp."},
+                {"name": "remote_setup.npm", "ok": True, "detail": "/opt/homebrew/bin/npm", "required": True, "title": "npm"},
+            ],
+            "print_fn": self.print_line,
+        }
+
+        result = self.mod.cmd_desktop_video_setup(
+            Namespace(
+                target="mac",
+                machine="blackbook",
+                check=True,
+                probe_host="blackbook",
+                skip_remotion_smoke=False,
+                video_audio="none",
+                video_audio_device=None,
+                json=True,
+            ),
+            **deps,
+        )
+
+        self.assertEqual(result, 1)
+        self.assertEqual(probed_hosts, ["blackbook"])
+        payload = json.loads(self.printed[0])
+        self.assertTrue(payload["setup_prerequisites"]["ok"])
+        self.assertFalse(payload["remote_setup_prerequisites"]["ok"])
+        self.assertEqual(payload["remote_setup_prerequisites"]["host"], "blackbook")
+        self.assertEqual(payload["remote_setup_prerequisites"]["remediations"][0]["check"], "remote_setup.pulp")
+        self.assertIsNone(payload["remote_setup_prerequisites"]["probe"])
+
+    def test_video_setup_text_reports_demo_matrix_blockers(self):
+        self.targets["mac"]["optional"] = {"video_capture": True}
+        checks = [
+            {"name": "receipt", "ok": True, "detail": "installed"},
+            {"name": "screencapture", "ok": True, "detail": "/usr/sbin/screencapture"},
+            {"name": "video_capture", "ok": True, "detail": "/repo/node_modules/ffmpeg-static/ffmpeg", "required": False},
+            {"name": "avfoundation_screen", "ok": True, "detail": "Capture screen 0 (3:)", "required": False},
+        ]
+        deps = {
+            "load_config_fn": self.config,
+            "resolve_desktop_target_fn": lambda _config, name: self.targets[name],
+            "desktop_doctor_checks_fn": lambda _config, _name: [dict(check) for check in checks],
+            "normalize_desktop_optional_config_fn": lambda optional: {"video_capture": bool((optional or {}).get("video_capture"))},
+            "video_proof_smoke_fn": lambda: {"ok": True, "detail": "smoke ok"},
+            "probe_macos_avfoundation_audio_fn": lambda _device: self.fail("audio probe should not run"),
+            "setup_prerequisite_checks_fn": self.setup_ok_checks,
+            "desktop_video_matrix_payload_fn": lambda **_kwargs: {
+                "kind": "desktop-video-proof-demo-matrix",
+                "checked": True,
+                "scenarios": [
+                    {
+                        "id": "audio-inspector-demo",
+                        "status": "ready",
+                        "declared_status": "ready",
+                    },
+                    {
+                        "id": "design-parity",
+                        "status": "blocked",
+                        "declared_status": "ready",
+                        "local_readiness": {
+                            "checks": [
+                                {
+                                    "name": "source_image",
+                                    "ok": False,
+                                    "required": True,
+                                    "detail": "missing planning/screenshots/reference.png",
+                                },
+                                {
+                                    "name": "optional_note",
+                                    "ok": False,
+                                    "required": False,
+                                    "detail": "ignored",
+                                },
+                            ],
+                        },
+                    },
+                ],
+            },
+            "print_fn": self.print_line,
+        }
+
+        result = self.mod.cmd_desktop_video_setup(
+            Namespace(target="mac", machine="blackbook", check=True, skip_remotion_smoke=False, video_audio="none", video_audio_device=None, json=False),
+            **deps,
+        )
+
+        self.assertEqual(result, 0)
+        self.assertIn("Setup prerequisites: PASS", self.printed)
+        self.assertIn("Demo matrix readiness:", self.printed)
+        self.assertIn("  - audio-inspector-demo: ready", self.printed)
+        self.assertIn("  - design-parity: blocked (declared: ready)", self.printed)
+        self.assertIn("      blocker: source_image: missing planning/screenshots/reference.png", self.printed)
+        self.assertFalse(any("optional_note" in line for line in self.printed))
+
+
+if __name__ == "__main__":
+    unittest.main()
